refact (java) : 수익률 계산 로직 변경 및 기타 수정

+++ src/main/java/lotto/domain/LottoResult.java
@@ -5,7 +5,7 @@ import java.util.Arrays;
 
 public class LottoResult {
 
-    private static final long LOTTO_AMOUNT = 1000;
+    private static final double LOTTO_AMOUNT = 1000;
     private int totalLottoCnt;
     private int totalPrize;
     private int fifthPlaceCnt;
